Restrict "qiime aldex2 aldex2" types (close #5)
Changes:
1. --p-denom is restricted to "all" or "iqlr"
2. --p-test is restricted to "t" or "glm"
3. Instead of specifying --p-condition, users now specify
   --m-metadata-file (like before) and --m-metadata-column (in place
   of --p-condition). This required a few code changes under the hood
   to get things working, but the external interface is the same as
   before (just with the renamed parameter name, of course).

Doing this through the QIIME 2 interface offers a few benefits. For
example, if you previously specified a --p-condition that wasn't in
the metadata file, this would have resulted in a confusing R error.
Now, you get an error from QIIME 2 saying there was an error in
finding the column name in the metadata.

I updated the tutorial accordingly (replacing --p-condition with
--m-metadata-column).

The automatic tests fail, but... they were already failing due to the
pearsonr() line (and they still fail at that line).